pipelines/generate_cogs/pyproject.toml
[project]
name = "vhm-generate-cogs"
version = "0.0.0"
requires-python = ">=3.11"

dependencies = [
  "python-dotenv",
  "rioxarray",
  "s3fs",
  "xarray",
  "zarr",
]

[project.optional-dependencies]
dev = [
  "ipykernel",
  "matplotlib",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
fix = true

[tool.ruff.lint]
extend-select = ["I", "RUF"]
